TS-29451 Added system test for testing the control server. Work in progress
