+++ packages/teamscale-coverage-collector/package.json
@@ -53,6 +53,7 @@
     "babel-jest": "^27.2.0",
     "esbuild": "^0.13.4",
     "jest": "^27.2.0",
+    "mockttp": "^3.1.0",
     "rimraf": "^3.0.2",
     "ts-jest": "^27.0.5",
     "ts-node": "^10.2.1",
